Ubicacion correcta de los archivo barrill de cada carpeta (modelos/controllers). Tambien se arreglaron las rutas de archivos mal puestos
