+++ src/controllers/index.controller.js
@@ -1,0 +1,15 @@
+import { create } from './users/create.controller.js';
+import { getAll } from './users/findAll.controller.js';
+import { getById } from './users/findById.controller.js';
+import { update } from './users/update.controller.js';
+import { updateStatus } from './users/updateStatus.controller.js';
+import { deleteUser } from './users/delete.controller.js';
+
+export {
+    create,
+    getAll,
+    getById,
+    update,
+    updateStatus,
+    deleteUser
+};
