--- src/controllers/users/index.controller.js
@@ -1,15 +1,0 @@
-import { create } from './create.controller.js';
-import { getAll } from './findAll.controller.js';
-import { getById } from './findById.controller.js';
-import { update } from './update.controller.js';
-import { updateStatus } from './updateStatus.controller.js';
-import { deleteUser } from './delete.controller.js';
-
-export {
-    create,
-    getAll,
-    getById,
-    update,
-    updateStatus,
-    deleteUser
-};
